Add new ModifyAnnotationProposal class and WebServlet quick fix
