+++ jakarta-eclipse/src/org/jakarta/codeAction/proposal/NewAnnotationProposal.java
@@ -22,7 +22,6 @@ import org.eclipse.jdt.core.dom.Annotation;
 import org.eclipse.jdt.core.dom.CompilationUnit;
 import org.eclipse.jdt.core.dom.FieldDeclaration;
 import org.eclipse.jdt.core.dom.IBinding;
-import org.eclipse.jdt.core.dom.ITypeBinding;
 import org.eclipse.jdt.core.dom.TypeDeclaration;
 import org.eclipse.jdt.core.dom.VariableDeclarationFragment;
 import org.eclipse.jdt.core.dom.rewrite.ASTRewrite;
@@ -32,86 +31,85 @@ import org.eclipse.jdt.internal.core.manipulation.dom.ASTResolving;
 import org.eclipse.jdt.internal.corext.codemanipulation.ContextSensitiveImportRewriteContext;
 import org.eclipse.lsp4j.CodeActionKind;
 
-
-/************************************************
- * Code action proposal for annotation
- * reused from https://github.com/eclipse/lsp4mp/blob/6f2d700a88a3262e39cc2ba04beedb429e162246/microprofile.jdt/org.eclipse.lsp4mp.jdt.core/src/main/java/org/eclipse/lsp4mp/jdt/core/java/corrections/proposal/NewAnnotationProposal.java
+/**
+ * Code action proposal for annotation reused from
+ * https://github.com/eclipse/lsp4mp/blob/6f2d700a88a3262e39cc2ba04beedb429e162246/microprofile.jdt/org.eclipse.lsp4mp.jdt.core/src/main/java/org/eclipse/lsp4mp/jdt/core/java/corrections/proposal/NewAnnotationProposal.java
  */
 public class NewAnnotationProposal extends ChangeCorrectionProposal {
 
-	private final CompilationUnit fInvocationNode;
-	private final IBinding fBinding;
+    private final CompilationUnit fInvocationNode;
+    private final IBinding fBinding;
 
-	private final String[] annotations;
+    private final String[] annotations;
 
-	public NewAnnotationProposal(String label, ICompilationUnit targetCU, CompilationUnit invocationNode,
-			IBinding binding, int relevance, String... annotations) {
-		super(label, CodeActionKind.QuickFix, targetCU, null, relevance);
-		fInvocationNode = invocationNode;
-		fBinding = binding;
-		this.annotations = annotations;
-	}
+    public NewAnnotationProposal(String label, ICompilationUnit targetCU, CompilationUnit invocationNode,
+            IBinding binding, int relevance, String... annotations) {
+        super(label, CodeActionKind.QuickFix, targetCU, null, relevance);
+        fInvocationNode = invocationNode;
+        fBinding = binding;
+        this.annotations = annotations;
+    }
 
-	@Override
-	protected ASTRewrite getRewrite() throws CoreException {
-		ASTNode declNode = null;
-		ASTNode boundNode = fInvocationNode.findDeclaringNode(fBinding);
-		CompilationUnit newRoot = fInvocationNode;
-		if (boundNode != null) {
-			declNode = boundNode; // is same CU
-		} else {
-			newRoot = ASTResolving.createQuickFixAST(getCompilationUnit(), null);
-			declNode = newRoot.findDeclaringNode(fBinding.getKey());
-		}
-		ImportRewrite imports = createImportRewrite(newRoot);
+    @Override
+    protected ASTRewrite getRewrite() throws CoreException {
+        ASTNode declNode = null;
+        ASTNode boundNode = fInvocationNode.findDeclaringNode(fBinding);
+        CompilationUnit newRoot = fInvocationNode;
+        if (boundNode != null) {
+            declNode = boundNode; // is same CU
+        } else {
+            newRoot = ASTResolving.createQuickFixAST(getCompilationUnit(), null);
+            declNode = newRoot.findDeclaringNode(fBinding.getKey());
+        }
+        ImportRewrite imports = createImportRewrite(newRoot);
 
-		boolean isField = declNode instanceof VariableDeclarationFragment;
-		if (isField) {
-			declNode = declNode.getParent();
-		}
-		if (declNode instanceof TypeDeclaration || isField) {
-			AST ast = declNode.getAST();
-			ASTRewrite rewrite = ASTRewrite.create(ast);
+        boolean isField = declNode instanceof VariableDeclarationFragment;
+        if (isField) {
+            declNode = declNode.getParent();
+        }
+        if (declNode instanceof TypeDeclaration || isField) {
+            AST ast = declNode.getAST();
+            ASTRewrite rewrite = ASTRewrite.create(ast);
 
-			ImportRewriteContext importRewriteContext = new ContextSensitiveImportRewriteContext(declNode, imports);
+            ImportRewriteContext importRewriteContext = new ContextSensitiveImportRewriteContext(declNode, imports);
 
-			for (String annotation : annotations) {
-				Annotation marker = ast.newMarkerAnnotation();
-				marker.setTypeName(ast.newName(imports.addImport(annotation, importRewriteContext))); // $NON-NLS-1$
-				rewrite.getListRewrite(declNode,
-						isField ? FieldDeclaration.MODIFIERS2_PROPERTY : TypeDeclaration.MODIFIERS2_PROPERTY)
-						.insertFirst(marker, null);
-			}
+            for (String annotation : annotations) {
+                Annotation marker = ast.newMarkerAnnotation();
+                marker.setTypeName(ast.newName(imports.addImport(annotation, importRewriteContext))); // $NON-NLS-1$
+                rewrite.getListRewrite(declNode,
+                        isField ? FieldDeclaration.MODIFIERS2_PROPERTY : TypeDeclaration.MODIFIERS2_PROPERTY)
+                        .insertFirst(marker, null);
+            }
 
-			return rewrite;
-		}
-		return null;
-	}
+            return rewrite;
+        }
+        return null;
+    }
 
-	/**
-	 * Returns the Compilation Unit node
-	 * 
-	 * @return the invocation node for the Compilation Unit
-	 */
-	protected CompilationUnit getInvocationNode() {
-		return this.fInvocationNode;
-	}
+    /**
+     * Returns the Compilation Unit node
+     * 
+     * @return the invocation node for the Compilation Unit
+     */
+    protected CompilationUnit getInvocationNode() {
+        return this.fInvocationNode;
+    }
 
-	/**
-	 * Returns the Binding object associated with the new annotation change
-	 * 
-	 * @return the binding object
-	 */
-	protected IBinding getBinding() {
-		return this.fBinding;
-	}
+    /**
+     * Returns the Binding object associated with the new annotation change
+     * 
+     * @return the binding object
+     */
+    protected IBinding getBinding() {
+        return this.fBinding;
+    }
 
-	/**
-	 * Returns the annotations list
-	 * 
-	 * @return the list of new annotations to add
-	 */
-	protected String[] getAnnotations() {
-		return this.annotations;
-	}
+    /**
+     * Returns the annotations list
+     * 
+     * @return the list of new annotations to add
+     */
+    protected String[] getAnnotations() {
+        return this.annotations;
+    }
 }
